New node-RED dashboard

+++ python/K9PythonController.py
@@ -191,13 +191,13 @@ class K9 :
       self.result = []
       self.left = self.leftMotor.getActualSpeed()
       self.right = self.rightMotor.getActualSpeed()
-      self.main_volt = (rc.ReadMainBatteryVoltage(rc_address)[1])/10
-      self.brain_volt = (rc.ReadLogicBatteryVoltage(rc_address)[1])/10
+      self.main_volt = float(rc.ReadMainBatteryVoltage(rc_address)[1])/10.0
+      self.brain_volt = float(rc.ReadLogicBatteryVoltage(rc_address)[1])/10.0
       self.currents = rc.ReadCurrents(rc_address)
       # Convert currents tuple in mAs to Amps
-      self.m1current = self.currents[1]/100
-      self.m2current = self.currents[2]/100
-      self.temp = (rc.ReadTemp(rc_address)[1])/10
+      self.m1current = float(self.currents[1]/100.0)
+      self.m2current = float(self.currents[2]/100.0)
+      self.temp = float(rc.ReadTemp(rc_address)[1])/10.0
       result = json.dumps({"type":"status","command":"update","left": self.left,"right": self.right,"lights": self.lights,"eyes": self.eyes,"hover": self.hover,"screen": self.screen, "motorctrl": self.motorctrl, "main_volt": self.main_volt, "brain_volt": self.brain_volt, "motor_l_amp": self.m1current, "motor_r_amp": self.m2current, "temp": self.temp }, skipkeys=False, ensure_ascii=True, check_circular=True, allow_nan=True, cls=None, indent=None, separators=(',', ': '), encoding="utf-8", default=None, sort_keys=False)
       return result
